requirements.txt
pyserial==3.5
Requests==2.32.3
